Split out subnet groups
